Add initial heist calculator (#5)
Create a custom public command that allows users to calculate how much
they need to heist in the **point heist mini-game** available in chat
with the following syntax:

`!heistcalc <current number of points> <desired goal>`

Adds an initial implementation of the heist calculator to test usage and
logic. In a future update, move this code to a module to encapsulate the
logic in its own class.

Tested with:
- `!heistcalc 962664 1000000`
- `!heistcalc 962664 500000`
- `!heistcalc 962664 5000000`
- `!heistcalc 1000 1450`
- `!heistcalc 1000 1`

These could be unit test cases.